WIP: Update Threshold ECDSA example
