--- examples/motoko_examples/threshold_ecdsa/canisters/motoko/main.mo
@@ -1,63 +1,0 @@
-import Cycles "mo:base/ExperimentalCycles";
-import Error "mo:base/Error";
-import Prim "mo:⛔";
-import Principal "mo:base/Principal";
-
-// Define the actor
-actor Assistant {
-
-  type Key = {
-    curve: Curve;
-    name: Text;
-  };
-
-  type Curve = {
-    #secp256k1;
-  };
-
-  // Only the ecdsa methods in the IC management canister is required here.
-  type IC = actor {
-    ecdsa_public_key : ({
-      canister_id : ?Principal;
-      derivation_path : [Blob];
-      key_id : Key;
-    }) -> async ({ public_key : Blob; chain_code : Blob; });
-    sign_with_ecdsa : ({
-      message_hash : Blob;
-      derivation_path : [Blob];
-      key_id : { curve: { #secp256k1; } ; name: Text };
-    }) -> async ({ signature : Blob });
-  };
-
-  let ic : IC = actor("aaaaa-aa");
-
-  public shared (msg) func public_key() : async { #Ok : { public_key: Blob }; #Err : Text } {
-    let caller = Principal.toBlob(msg.caller);
-    try {
-      let { public_key } = await ic.ecdsa_public_key({
-          canister_id = null;
-          derivation_path = [ caller ];
-          key_id = { curve = #secp256k1; name = "dfx_test_key" };
-      });
-      #Ok({ public_key })
-    } catch (err) {
-      #Err(Error.message(err))
-    }
-  };
-
-  public shared (msg) func sign(message_hash: Blob) : async { #Ok : { signature: Blob };  #Err : Text } {
-    assert(message_hash.size() == 32);
-    let caller = Principal.toBlob(msg.caller);
-    try {
-      Cycles.add(10_000_000_000);
-      let { signature } = await ic.sign_with_ecdsa({
-          message_hash;
-          derivation_path = [ caller ];
-          key_id = { curve = #secp256k1; name = "dfx_test_key" };
-      });
-      #Ok({ signature })
-    } catch (err) {
-      #Err(Error.message(err))
-    }
-  };
-}
