--- examples/motoko_examples/threshold_ecdsa/canisters/rust/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "rust"
-version = "0.0.0"
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-ic-cdk = "0.5.2"
-ic-cdk-macros = "0.5.2"
-candid = "0.7.14"
-serde = "1"
