--- examples/motoko_examples/threshold_ecdsa/canisters/rust/src/lib.did
@@ -1,4 +1,0 @@
-service : {
-  public_key : () -> (variant { Ok: record { public_key: blob; }; Err: text });
-  sign : (blob) -> (variant { Ok: record { signature: blob; }; Err: text });
-}
