+++ examples/motoko_examples/threshold_ecdsa/package.json
@@ -8,8 +8,8 @@
         "encode-utf8": "2.0.0"
     },
     "devDependencies": {
-        "@dfinity/agent": "^0.12.1",
-        "ts-node": "10.9.1",
-        "typescript": "^4.7.4"
+        "@dfinity/agent": "^0.19.2",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.2.2"
     }
 }
