+++ examples/motoko_examples/threshold_ecdsa/src/index.did
