+++ examples/motoko_examples/threshold_ecdsa/test/tests.ts
@@ -1,5 +1,5 @@
 import { Test } from 'azle/test';
-import { _SERVICE } from '../dfx_generated/azle/azle.did';
+import { _SERVICE } from './dfx_generated/threshold_ecdsa/threshold_ecdsa.did';
 import { ActorSubclass } from '@dfinity/agent';
 
 export function getTests(tEcdsaCanister: ActorSubclass<_SERVICE>): Test[] {
